webjobs refactor - addressing PR comments
