+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/QueueTriggers/GenerateAllVacancyAnalyticsQueueTrigger.cs
@@ -10,17 +10,19 @@ using Esfa.Recruit.Vacancies.Jobs.Configuration;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.QueueTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.QueueTriggers
 {
     public class GenerateAllVacancyAnalyticsQueueTrigger
     {
-        private readonly ILogger<GenerateVacancyAnalyticsQueueTrigger> _logger;
+        private readonly ILogger<GenerateVacancyAnalyticsSummaryQueueTrigger> _logger;
         private readonly RecruitWebJobsSystemConfiguration _jobsConfig;
-        private readonly IQueue _queue;
+        private readonly IQueueService _queue;
         private readonly IVacancyQuery _vacancyQuery;
 
-        public GenerateAllVacancyAnalyticsQueueTrigger(ILogger<GenerateVacancyAnalyticsQueueTrigger> logger, RecruitWebJobsSystemConfiguration jobsConfig,
-            IQueue queue, IVacancyQuery vacancyQuery)
+        private string JobName => GetType().Name;
+
+        public GenerateAllVacancyAnalyticsQueueTrigger(ILogger<GenerateVacancyAnalyticsSummaryQueueTrigger> logger, RecruitWebJobsSystemConfiguration jobsConfig,
+            IQueueService queue, IVacancyQuery vacancyQuery)
         {
             _logger = logger;
             _jobsConfig = jobsConfig;
@@ -30,9 +32,9 @@ namespace Esfa.Recruit.Vacancies.Jobs.QueueTriggers
 
         public async Task GenerateAllVacancyAnalyticsAsync([QueueTrigger(QueueNames.GenerateAllVacancyAnalyticsSummariesQueueName, Connection = "QueueStorage")] string message, TextWriter log)
         {
-            if (_jobsConfig.DisabledJobs.Contains(this.GetType().Name))
+            if (_jobsConfig.DisabledJobs.Contains(JobName))
             {
-                _logger.LogDebug($"{this.GetType().Name} is disabled, skipping ...");
+                _logger.LogDebug($"{JobName} is disabled, skipping ...");
                 return;
             }
