+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/QueueTriggers/GenerateReportQueueTrigger.cs
@@ -9,7 +9,7 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 
-namespace Esfa.Recruit.Vacancies.Jobs.QueueTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.QueueTriggers
 {
     public class GenerateReportQueueTrigger
     {
