+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/QueueTriggers/SpikeQueueTrigger.cs
@@ -7,13 +7,15 @@ using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
 #if DEBUG
-namespace Esfa.Recruit.Vacancies.Jobs.QueueTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.QueueTriggers
 {
     public class SpikeQueueTrigger
     {
         private readonly ILogger<SpikeQueueTrigger> _logger;
         private readonly RecruitWebJobsSystemConfiguration _jobsConfig;
         private readonly IVacancySummariesProvider _query;
+
+        private string JobName => GetType().Name;
 
         public SpikeQueueTrigger(ILogger<SpikeQueueTrigger> logger, RecruitWebJobsSystemConfiguration jobsConfig, IVacancySummariesProvider query)
         {
@@ -24,6 +26,12 @@ namespace Esfa.Recruit.Vacancies.Jobs.QueueTriggers
 
         public async Task SpikeAsync([QueueTrigger("test-queue", Connection = "QueueStorage")] string message, TextWriter log)
         {
+            if (_jobsConfig.DisabledJobs.Contains(JobName))
+            {
+                _logger.LogDebug($"{JobName} is disabled, skipping ...");
+                return;
+            }
+
             var employerAccountId = Environment.GetEnvironmentVariable("EmployerAccount");
             var result = await _query.GetEmployerOwnedVacancySummariesByEmployerAccountAsync(employerAccountId);
         }
