+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/TimerTriggers/GenerateBlockedEmployersTimeTrigger.cs
@@ -6,26 +6,26 @@ using Esfa.Recruit.Vacancies.Client.Application.Queues.Messages;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.TimerTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.TimerTriggers
 {
-    public class VacancyStatusTimeTrigger
+    public class GenerateBlockedEmployersTimeTrigger
     {
-        private readonly ILogger<VacancyStatusTimeTrigger> _logger;
-        private readonly IQueue _queue;
+        private readonly ILogger<GenerateBlockedEmployersTimeTrigger> _logger;
+        private readonly IQueueService _queue;
         private readonly ITimeProvider _timeProvider;
 
-        public VacancyStatusTimeTrigger(ILogger<VacancyStatusTimeTrigger> logger, IQueue queue, ITimeProvider timeProvider)
+        public GenerateBlockedEmployersTimeTrigger(ILogger<GenerateBlockedEmployersTimeTrigger> logger, IQueueService queue, ITimeProvider timeProvider)
         {
             _logger = logger;
             _queue = queue;
             _timeProvider = timeProvider;
         }
 
-        public Task VacancyStatusAsync([TimerTrigger(Schedules.MidnightDaily)] TimerInfo timerInfo, TextWriter log)
+        public Task UpdateBlockedEmployersAsync([TimerTrigger(Schedules.Hourly)] TimerInfo timerInfo, TextWriter log)
         {
             _logger.LogInformation($"Timer trigger {this.GetType().Name} fired");
 
-            var message = new VacancyStatusQueueMessage
+            var message = new UpdateBlockedEmployersQueueMessage 
             {
                 CreatedByScheduleDate = _timeProvider.Now
             };
