+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/TimerTriggers/UpdateApprenticeshipProgrammesTimerTrigger.cs
@@ -6,26 +6,26 @@ using Esfa.Recruit.Vacancies.Client.Application.Queues.Messages;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.TimerTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.TimerTriggers
 {
-    public class UpdateQaDashboardTimeTrigger
+    public class UpdateApprenticeshipProgrammesTimerTrigger
     {
-        private readonly ILogger<UpdateQaDashboardTimeTrigger> _logger;
-        private readonly IQueue _queue;
+        private readonly ILogger<UpdateApprenticeshipProgrammesTimerTrigger> _logger;
+        private readonly IQueueService _queue;
         private readonly ITimeProvider _timeProvider;
-
-        public UpdateQaDashboardTimeTrigger(ILogger<UpdateQaDashboardTimeTrigger> logger, IQueue queue, ITimeProvider timeProvider)
+        
+        public UpdateApprenticeshipProgrammesTimerTrigger(ILogger<UpdateApprenticeshipProgrammesTimerTrigger> logger, IQueueService queue, ITimeProvider timeProvider)
         {
             _logger = logger;
             _queue = queue;
             _timeProvider = timeProvider;
         }
 
-        public Task UpdateQaDashboardAsync([TimerTrigger(Schedules.EveryFiveMinutes)] TimerInfo timerInfo, TextWriter log)
+        public Task UpdateApprenticeshipProgrammesAsync([TimerTrigger(Schedules.FourAmDaily)] TimerInfo timerInfo, TextWriter log)
         {
             _logger.LogInformation($"Timer trigger {this.GetType().Name} fired");
 
-            var message = new UpdateQaDashboardQueueMessage
+            var message = new UpdateApprenticeshipProgrammesQueueMessage
             {
                 CreatedByScheduleDate = _timeProvider.Now
             };
