+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/TimerTriggers/UpdateBankHolidaysTimerTrigger.cs
@@ -6,26 +6,26 @@ using Esfa.Recruit.Vacancies.Client.Application.Queues.Messages;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.TimerTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.TimerTriggers
 {
-    public class DeleteReportsTimeTrigger
+    public class UpdateBankHolidaysTimerTrigger
     {
-        private readonly ILogger<DeleteReportsTimeTrigger> _logger;
-        private readonly IQueue _queue;
+        private readonly ILogger<UpdateBankHolidaysTimerTrigger> _logger;
+        private readonly IQueueService _queue;
         private readonly ITimeProvider _timeProvider;
 
-        public DeleteReportsTimeTrigger(ILogger<DeleteReportsTimeTrigger> logger, IQueue queue, ITimeProvider timeProvider)
+        public UpdateBankHolidaysTimerTrigger(ILogger<UpdateBankHolidaysTimerTrigger> logger, IQueueService queue, ITimeProvider timeProvider)
         {
             _logger = logger;
             _queue = queue;
             _timeProvider = timeProvider;
         }
 
-        public Task DeleteReportsAsync([TimerTrigger(Schedules.MidnightDaily)] TimerInfo timerInfo, TextWriter log)
+        public Task UpdateBankHolidaysAsync([TimerTrigger(Schedules.FourAmDaily)] TimerInfo timerInfo, TextWriter log)
         {
             _logger.LogInformation($"Timer trigger {this.GetType().Name} fired");
 
-            var message = new DeleteReportsQueueMessage 
+            var message = new UpdateBankHolidaysQueueMessage
             {
                 CreatedByScheduleDate = _timeProvider.Now
             };
