+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/TimerTriggers/UpdateQaDashboardTimeTrigger.cs
@@ -6,26 +6,26 @@ using Esfa.Recruit.Vacancies.Client.Application.Queues.Messages;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.TimerTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.TimerTriggers
 {
-    public class GenerateBlockedEmployersTimeTrigger
+    public class UpdateQaDashboardTimeTrigger
     {
-        private readonly ILogger<GenerateBlockedEmployersTimeTrigger> _logger;
-        private readonly IQueue _queue;
+        private readonly ILogger<UpdateQaDashboardTimeTrigger> _logger;
+        private readonly IQueueService _queue;
         private readonly ITimeProvider _timeProvider;
 
-        public GenerateBlockedEmployersTimeTrigger(ILogger<GenerateBlockedEmployersTimeTrigger> logger, IQueue queue, ITimeProvider timeProvider)
+        public UpdateQaDashboardTimeTrigger(ILogger<UpdateQaDashboardTimeTrigger> logger, IQueueService queue, ITimeProvider timeProvider)
         {
             _logger = logger;
             _queue = queue;
             _timeProvider = timeProvider;
         }
 
-        public Task UpdateBlockedEmployersAsync([TimerTrigger(Schedules.Hourly)] TimerInfo timerInfo, TextWriter log)
+        public Task UpdateQaDashboardAsync([TimerTrigger(Schedules.EveryFiveMinutes)] TimerInfo timerInfo, TextWriter log)
         {
             _logger.LogInformation($"Timer trigger {this.GetType().Name} fired");
 
-            var message = new UpdateBlockedEmployersQueueMessage 
+            var message = new UpdateQaDashboardQueueMessage
             {
                 CreatedByScheduleDate = _timeProvider.Now
             };
