+++ src/Jobs/Recruit.Vacancies.Jobs/Triggers/TimerTriggers/VacancyStatusTimeTrigger.cs
@@ -6,26 +6,26 @@ using Esfa.Recruit.Vacancies.Client.Application.Queues.Messages;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Extensions.Logging;
 
-namespace Esfa.Recruit.Vacancies.Jobs.TimerTriggers
+namespace Esfa.Recruit.Vacancies.Jobs.Triggers.TimerTriggers
 {
-    public class UpdateBankHolidaysTimerTrigger
+    public class VacancyStatusTimeTrigger
     {
-        private readonly ILogger<UpdateBankHolidaysTimerTrigger> _logger;
-        private readonly IQueue _queue;
+        private readonly ILogger<VacancyStatusTimeTrigger> _logger;
+        private readonly IQueueService _queue;
         private readonly ITimeProvider _timeProvider;
 
-        public UpdateBankHolidaysTimerTrigger(ILogger<UpdateBankHolidaysTimerTrigger> logger, IQueue queue, ITimeProvider timeProvider)
+        public VacancyStatusTimeTrigger(ILogger<VacancyStatusTimeTrigger> logger, IQueueService queue, ITimeProvider timeProvider)
         {
             _logger = logger;
             _queue = queue;
             _timeProvider = timeProvider;
         }
 
-        public Task UpdateBankHolidaysAsync([TimerTrigger(Schedules.FourAmDaily)] TimerInfo timerInfo, TextWriter log)
+        public Task VacancyStatusAsync([TimerTrigger(Schedules.MidnightDaily)] TimerInfo timerInfo, TextWriter log)
         {
             _logger.LogInformation($"Timer trigger {this.GetType().Name} fired");
 
-            var message = new UpdateBankHolidaysQueueMessage
+            var message = new VacancyStatusQueueMessage
             {
                 CreatedByScheduleDate = _timeProvider.Now
             };
